demo/modal_client/requirements.txt
requests>=2.31.0
numpy
opencv-python
Pillow
